[interpreter] clean up exported symbols and header organization
