+++ mono/mini/driver.c
@@ -2067,7 +2067,7 @@ mono_main (int argc, char* argv[])
 	case DO_REGRESSION:
 #ifdef ENABLE_INTERPRETER
 		if (mono_use_interpreter) {
-			if (interp_regression_list (2, argc -i, argv + i)) {
+			if (mono_interp_regression_list (2, argc -i, argv + i)) {
 				g_print ("Regression ERRORS!\n");
 				// mini_cleanup (domain);
 				return 1;
